product filtering from Elasticsearch is now moved to shopsys/framework
