+++ packages/framework/src/Model/Product/Flag/FlagRepository.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Model\Product\Flag;
 
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Query\Expr\Join;
+use Shopsys\FrameworkBundle\Component\Doctrine\OrderByCollationHelper;
 use Shopsys\FrameworkBundle\Model\Product\Flag\Exception\FlagNotFoundException;
 
 class FlagRepository
@@ -90,5 +92,24 @@ class FlagRepository
     public function getByUuids(array $uuids): array
     {
         return $this->getFlagRepository()->findBy(['uuid' => $uuids]);
+    }
+
+    /**
+     * @param int[] $flagsIds
+     * @param string $locale
+     * @return \Shopsys\FrameworkBundle\Model\Product\Flag\Flag[]
+     */
+    public function getVisibleFlagsByIds(array $flagsIds, string $locale): array
+    {
+        $flagsQueryBuilder = $this->getFlagRepository()->createQueryBuilder('f')
+            ->select('f, ft')
+            ->join('f.translations', 'ft', Join::WITH, 'ft.locale = :locale')
+            ->where('f.id IN (:flagsIds)')
+            ->andWhere('f.visible = true')
+            ->orderBy(OrderByCollationHelper::createOrderByForLocale('ft.name', $locale), 'asc')
+            ->setParameter('flagsIds', $flagsIds)
+            ->setParameter('locale', $locale);
+
+        return $flagsQueryBuilder->getQuery()->getResult();
     }
 }
