+++ project-base/app/src/Model/Product/Brand/BrandRepository.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace App\Model\Product\Brand;
 
-use App\Component\Doctrine\OrderByCollationHelper;
-use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\QueryBuilder;
-use Shopsys\FrameworkBundle\Component\Domain\Domain;
+use Shopsys\FrameworkBundle\Component\Doctrine\OrderByCollationHelper;
 use Shopsys\FrameworkBundle\Component\String\DatabaseSearching;
 use Shopsys\FrameworkBundle\Model\Product\Brand\BrandRepository as BaseBrandRepository;
 
@@ -16,20 +14,10 @@ use Shopsys\FrameworkBundle\Model\Product\Brand\BrandRepository as BaseBrandRepo
  * @method \App\Model\Product\Brand\Brand[] getAll()
  * @method \App\Model\Product\Brand\Brand getOneByUuid(string $uuid)
  * @method \App\Model\Product\Brand\Brand[] getByUuids(string[] $uuids)
+ * @method \App\Model\Product\Brand\Brand[] getBrandsByIds(int[] $brandsIds)
  */
 class BrandRepository extends BaseBrandRepository
 {
-    /**
-     * @param \Doctrine\ORM\EntityManagerInterface $entityManager
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
-     */
-    public function __construct(
-        EntityManagerInterface $entityManager,
-        private Domain $domain,
-    ) {
-        parent::__construct($entityManager);
-    }
-
     /**
      * @param string $searchText
      * @return array
@@ -56,20 +44,5 @@ class BrandRepository extends BaseBrandRepository
         $queryBuilder->setParameter('searchText', DatabaseSearching::getFullTextLikeSearchString($searchText));
 
         return $queryBuilder;
-    }
-
-    /**
-     * @param int[] $brandsIds
-     * @return \App\Model\Product\Brand\Brand[]
-     */
-    public function getBrandsByIds(array $brandsIds): array
-    {
-        $brandsQueryBuilder = $this->getBrandRepository()->createQueryBuilder('b')
-            ->select('b')
-            ->where('b.id IN (:brandIds)')
-            ->setParameter('brandIds', $brandsIds)
-            ->orderBy(OrderByCollationHelper::createOrderByForLocale('b.name', $this->domain->getLocale()), 'asc');
-
-        return $brandsQueryBuilder->getQuery()->getResult();
     }
 }
