+++ project-base/app/src/Model/Product/Parameter/ParameterFacade.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace App\Model\Product\Parameter;
 
 use App\Component\UploadedFile\UploadedFileFacade;
-use App\Model\Category\Category;
 use App\Model\CategorySeo\ReadyCategorySeoMixFacade;
 use Doctrine\ORM\EntityManagerInterface;
-use Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice;
 use Shopsys\FrameworkBundle\Model\Category\CategoryParameterRepository;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterFacade as BaseParameterFacade;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterFactoryInterface;
@@ -29,6 +27,7 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
  * @method \App\Model\Product\Parameter\ParameterValue getParameterValueByUuid(string $uuid)
  * @method \App\Model\Product\Parameter\Parameter[] getParametersByUuids(string[] $uuids)
  * @method \App\Model\Product\Parameter\ParameterValue[] getParameterValuesByUuids(string[] $uuids)
+ * @method int[] getParametersIdsSortedByPositionFilteredByCategory(\App\Model\Category\Category $category)
  */
 class ParameterFacade extends BaseParameterFacade
 {
@@ -37,24 +36,25 @@ class ParameterFacade extends BaseParameterFacade
      * @param \App\Model\Product\Parameter\ParameterRepository $parameterRepository
      * @param \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterFactory $parameterFactory
      * @param \Symfony\Component\EventDispatcher\EventDispatcherInterface $eventDispatcher
+     * @param \Shopsys\FrameworkBundle\Model\Category\CategoryParameterRepository $categoryParameterRepository
      * @param \App\Model\CategorySeo\ReadyCategorySeoMixFacade $readyCategorySeoMixFacade
      * @param \App\Component\UploadedFile\UploadedFileFacade $uploadedFileFacade
-     * @param \App\Model\Category\CategoryParameterRepository $categoryParameterRepository
      */
     public function __construct(
         EntityManagerInterface $em,
         ParameterRepository $parameterRepository,
         ParameterFactoryInterface $parameterFactory,
         EventDispatcherInterface $eventDispatcher,
+        CategoryParameterRepository $categoryParameterRepository,
         private readonly ReadyCategorySeoMixFacade $readyCategorySeoMixFacade,
         private readonly UploadedFileFacade $uploadedFileFacade,
-        private readonly CategoryParameterRepository $categoryParameterRepository,
     ) {
         parent::__construct(
             $em,
             $parameterRepository,
             $parameterFactory,
             $eventDispatcher,
+            $categoryParameterRepository,
         );
     }
 
@@ -144,52 +144,5 @@ class ParameterFacade extends BaseParameterFacade
         $this->em->flush();
 
         return $parameterValue;
-    }
-
-    /**
-     * @param int[][] $parameterValueIdsIndexedByParameterId
-     * @param string $locale
-     * @return \Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[]
-     */
-    public function getParameterFilterChoicesByIds(array $parameterValueIdsIndexedByParameterId, string $locale): array
-    {
-        $parameterValueIds = array_reduce($parameterValueIdsIndexedByParameterId, 'array_merge', []);
-        $allParameters = $this->parameterRepository->getVisibleParametersByIds(
-            array_keys($parameterValueIdsIndexedByParameterId),
-            $locale,
-        );
-        $allParameterValues = $this->parameterRepository->getParameterValuesByIds($parameterValueIds);
-
-        $parameterFilterChoices = [];
-
-        foreach ($allParameters as $parameter) {
-            $valueIdsForParameter = $parameterValueIdsIndexedByParameterId[$parameter->getId()];
-            $parameterValues = array_intersect_key($allParameterValues, array_flip($valueIdsForParameter));
-
-            uasort($parameterValues, function (ParameterValue $first, ParameterValue $second) {
-                return strcmp($first->getText(), $second->getText());
-            });
-
-            $parameterFilterChoices[] = new ParameterFilterChoice(
-                $parameter,
-                $parameterValues,
-            );
-        }
-
-        return $parameterFilterChoices;
-    }
-
-    /**
-     * @param \App\Model\Category\Category $category
-     * @return int[]
-     */
-    public function getParametersIdsSortedByPositionFilteredByCategory(Category $category): array
-    {
-        return array_map(
-            function ($categoryParameter) {
-                return $categoryParameter->getParameter()->getId();
-            },
-            $this->categoryParameterRepository->getCategoryParametersByCategorySortedByPosition($category),
-        );
     }
 }
